Move text stuff into text_formatters
