+++ text/homophones.py
@@ -6,7 +6,7 @@ from talon.voice import Word, Key, Context, Str, press
 from talon import canvas
 from talon.skia import Rect
 
-from ..misc.std import parse_word
+from ..utils import parse_word
 import os
 
 ########################################################################
